size_t instead of int

+++ corepurity_core/src/graphics/buffers/vertexarray.cpp
@@ -10,7 +10,7 @@ namespace corepurity_core { namespace graphics {
 
     VertexArray::~VertexArray()
     {
-        for (int i = 0; i < m_Buffers.size(); i++)
+        for (size_t i = 0; i < m_Buffers.size(); i++)
         {
             delete m_Buffers[i];
         }
